Add email error mapping and move mapper to common
